kill now fails gracefully

+++ vpn.sh
@@ -22,10 +22,29 @@ display_help() {
 }
 
 kill_ssh() {
-    ssh newton.ist.ucf.edu -O exit
-    ssh stokes.ist.ucf.edu -O exit
-    sudo kill $(cat $PIDFILE)
+    echo "Shutting down ARCC VPN components"
+
+    socket_newton=$(ssh -G newton.ist.ucf.edu | awk '/^controlpath / {print $2}')
+    socket_stokes=$(ssh -G stokes.ist.ucf.edu | awk '/^controlpath / {print $2}')
+
+    if [ -S "$socket_newton" ]; then
+        ssh -S "$socket_newton" newton.ist.ucf.edu -O exit
+        echo "  Terminating shared connection to newton.ist.ucf.edu"
+    fi
+
+    if [ -S "$socket_stokes" ]; then
+        ssh -S "$socket_stokes" stokes.ist.ucf.edu -O exit
+        echo "  Terminating shared connection to newton.ist.ucf.edu"
+    fi
+
+    if [ -f "$PIDFILE" ]; then
+        sudo kill $(cat $PIDFILE)
+        echo "  Terminated ssh process"
+    fi
+    echo   "ARCC VPN is no longer running"
+    exit 0
 }
+
 
 show_bookmarklet(){
     echo "Add a bookmark in your with the following URL: (copied to your clipboard)"
@@ -56,6 +75,15 @@ while true; do
 done
 
 
+display_vpn_login_instructions() {
+    echo "=====================INSTRUCTIONS==============================="
+    echo -e "Follow the prompts at ${BLUE}https://secure.vpn.ucf.edu/${NC} to authenticate with the UCF VPN."
+    echo "Then use the bookmarklet to copy your authentication Cookie 🍪."
+    echo "Bookmarklet command:"
+    echo "    javascript:window.alert(document.cookie.match(/webvpn=(.*);/)[1])"
+    echo "=====================++++++++++++==============================="
+}
+
 display_instructions() {
     echo "=====================INSTRUCTIONS==============================="
     echo $salloc_command | xclip -selection clipboard && echo "📋 Interactive session command copied to your clipboard :
@@ -77,11 +105,14 @@ Next step:
         --time=01:00:00
  3. DO NOT close this terminal; salloc starts an INTERACTIVE session!
 -----------------------------------------------------------------------
-Shutdown: use `$0 --kill` or...
- 1. Clean up port forwarding with:
-    ssh $server_name -O exit
- 2. Terminate VPN connection with:
-    sudo kill $(cat $PIDFILE)
+Shutdown: There are two methods to shutdown your connection.
+ Automatic:
+   1. run \`$0 --kill\`
+ Manual: 
+   1. Clean up port forwarding with:
+      ssh $server_name -O exit
+   2. Terminate VPN connection with:
+      sudo kill $(cat $PIDFILE)
 -----------------------------------------------------------------------
 EOF
 }
@@ -91,17 +122,13 @@ test $UID -ne 0 && echo 👑 Must be root  && exit 2
 BLUE="\033[1;34m"
 NC="\033[0m"
 
-echo "=====================INSTRUCTIONS==============================="
-echo -e "Follow the prompts at ${BLUE}https://secure.vpn.ucf.edu/${NC} to authenticate with the UCF VPN."
-echo "Then use the bookmarklet to copy your authentication cookie."
-echo "Bookmarklet command:"
-echo "    javascript:window.alert(document.cookie.match(/webvpn=(.*);/)[1])"
-echo "=====================++++++++++++==============================="
 
 test -f $PIDFILE && pgrep -aF $PIDFILE > /dev/null && echo 👍 openconnect already running  && display_instructions && exit 3
 
+display_vpn_login_instructions
+
 # javascript:window.alert(document.cookie.match(/webvpn=(.*?);/)[1])
-printf "Cookie 🍪: "
+printf "Paste the Cookie 🍪: "
 read COOKIE
 
 test -z "$COOKIE" && echo Must offer cookie 👹🍪 && exit 4
